+++ packages/db/src/lib/granule.ts
@@ -175,7 +175,7 @@ export const getApiGranuleExecutionCumulusIds = async (
  * @param {Object} searchParams
  * @param {string | Array<string>} [searchParams.collectionIds] - Collection ID
  * @param {string | Array<string>} [searchParams.granuleIds] - array of granule IDs
- * @param {string} [searchParams.providerName] - Provider name
+ * @param {string} [searchParams.providerNames] - Provider names
  * @param {UpdatedAtRange} [searchParams.updatedAtRange] - Date range for updated_at column
  * @param {string} [searchParams.status] - Granule status to search by
  * @param {string | Array<string>} [sortByFields] - Field(s) to sort by
@@ -249,4 +249,23 @@ export const getGranulesByApiPropertiesQuery = (
     .groupBy(`${granulesTable}.cumulus_id`)
     .groupBy(`${collectionsTable}.cumulus_id`)
     .groupBy(`${providersTable}.cumulus_id`);
+};
+
+/**
+ * Get granules from table where granule_id matches provided granuleId
+ *
+ * @param {Knex | Knex.Transaction} knexOrTransaction - DB client or transaction
+ * @param {string} granuleId - Granule ID
+ * @returns {Promise<Array<PostgresGranuleRecord>} The returned list of records
+ */
+export const getGranulesByGranuleId = async (
+  knexOrTransaction: Knex | Knex.Transaction,
+  granuleId: string
+): Promise<Array<PostgresGranuleRecord>> => {
+  const {
+    granules: granulesTable,
+  } = TableNames;
+  const records: PostgresGranuleRecord[] = await knexOrTransaction(granulesTable)
+    .where({ granule_id: granuleId });
+  return records;
 };